requirements.txt
freetype-py~=2.4.0
opencv-python~=4.7.0.72
numpy~=1.23.5
argparse~=1.4.0
pathlib~=1.0.1